code/Backend/ComposeGenerator/requirements.txt
PyYAML==6.0.1
marshmallow==3.21.1
